requirements.txt
# FLYTAU Dependencies
# Python 3.9+ required

# Web Framework
flask>=2.3.0

# MySQL Database Connector
mysql-connector-python>=8.0.0

# Password Hashing
bcrypt>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Development Tools (optional)
pytest>=7.0.0
pytest-flask>=1.2.0
